# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Knowledge Platform API",
        "description": "Read models, search and knowledge-graph APIs for the knowledge\npublishing platform.\n\nThe backend returns JSON read models only. Article bodies arrive as a\nstructured block tree rather than HTML: presentation belongs to the\nfrontend, and keeping markup out of the API is also what keeps the\nfrontend from becoming a sanitisation boundary.\n\nPublic read endpoints are anonymous. Everything under /api/studio\nrequires the AUTHOR role. Webhook endpoints are anonymous at the HTTP\nlayer and authenticated by provider signature.\n",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/webhooks/sources/{repositoryId}": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Accept a provider webhook for one connected repository",
                "operationId": "receive",
                "parameters": [
                    {
                        "name": "repositoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/studio/articles/{slug}/archive": {
            "post": {
                "tags": [
                    "Studio: Articles"
                ],
                "summary": "Archive an article, retaining it so inbound links are not orphaned",
                "operationId": "archive",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/topics/{slug}": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "The topic landing page read model",
                "operationId": "topic",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopicReadModel"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tags/{slug}": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "The tag landing page read model",
                "operationId": "tag",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagReadModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/studio/ingestions": {
            "get": {
                "tags": [
                    "Studio: Ingestion"
                ],
                "summary": "Recent ingestion runs, without diagnostics",
                "operationId": "recent",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "repositoryId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IngestionRunResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/studio/ingestions/{id}": {
            "get": {
                "tags": [
                    "Studio: Ingestion"
                ],
                "summary": "One ingestion run, with every diagnostic it produced",
                "operationId": "get_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestionRunResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/routes/resolve": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "Resolve a public URL path to what lives at it",
                "description": "The backend owns URL structure, so the frontend asks what a path is rather than\ninferring it from path shape. That is what lets an article be renamed without the\nfrontend needing to know.\n",
                "operationId": "resolveRoute",
                "parameters": [
                    {
                        "name": "path",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RouteTarget"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/authors/{slug}/page": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "The author page read model",
                "operationId": "author",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthorReadModel"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "The article archive, newest first",
                "operationId": "articles",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseArticleReference"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{slug}": {
            "get": {
                "tags": [
                    "Delivery"
                ],
                "summary": "The complete article page read model",
                "description": "Includes body blocks, outline, author, taxonomy, related articles, backlinks,\nbreadcrumbs, previous/next navigation and a bounded graph neighbourhood --\neverything a normal article page needs, with no follow-up calls.\n",
                "operationId": "article",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleReadModel"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ArticleResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "publicationState": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "authorId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "topics": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "wordCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "readingTimeMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "revisionNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sourcePath": {
                        "type": "string"
                    },
                    "sourceRevision": {
                        "type": "string"
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ArticleReference": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "summary": {
                        "type": "string"
                    },
                    "authorName": {
                        "type": "string"
                    }
                }
            },
            "TopicReadModel": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "articles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleReference"
                        }
                    },
                    "articleCount": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "TagReadModel": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "articles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleReference"
                        }
                    },
                    "articleCount": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Diagnostic": {
                "type": "object",
                "properties": {
                    "sourcePath": {
                        "type": "string"
                    },
                    "articleSlug": {
                        "type": "string"
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "INFO",
                            "WARNING",
                            "ERROR"
                        ]
                    },
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "IngestionRunResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "correlationId": {
                        "type": "string"
                    },
                    "repositoryId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "trigger": {
                        "type": "string",
                        "enum": [
                            "WEBHOOK",
                            "MANUAL",
                            "SCHEDULED"
                        ]
                    },
                    "state": {
                        "type": "string",
                        "enum": [
                            "RUNNING",
                            "SUCCEEDED",
                            "PARTIALLY_SUCCEEDED",
                            "FAILED"
                        ]
                    },
                    "requestedRevision": {
                        "type": "string"
                    },
                    "documentsTotal": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "documentsSucceeded": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "documentsSkipped": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "documentsFailed": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "startedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "finishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "diagnostics": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Diagnostic"
                        }
                    }
                }
            },
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "RouteTarget": {
                "type": "object",
                "properties": {
                    "targetType": {
                        "type": "string",
                        "enum": [
                            "ARTICLE",
                            "AUTHOR",
                            "TOPIC",
                            "TAG"
                        ]
                    },
                    "targetId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "targetSlug": {
                        "type": "string"
                    }
                }
            },
...
            "AuthorReadModel": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "biography": {
                        "type": "string"
                    },
                    "avatarUrl": {
                        "type": "string"
                    },
                    "articles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleReference"
                        }
                    },
                    "topics": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaxonomyReference"
                        }
                    },
                    "articleCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "joinedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "TaxonomyReference": {
                "type": "object",
                "properties": {
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "PageResponseArticleReference": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleReference"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}